# docs/swagger.yaml
basePath: /
definitions:
  domain.Authentication:
    properties:
      token:
        type: string
    type: object
  domain.Login:
    properties:
      password:
        maxLength: 12
        minLength: 3
        type: string
      username:
        maxLength: 10
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  domain.User:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Auth Input
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/domain.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Authentication'
      summary: Login
      tags:
      - Auth
    put:
      consumes:
      - application/json
      description: Signup
      parameters:
      - description: Sign up Input
        in: body
        name: signup
        required: true
        schema:
          $ref: '#/definitions/domain.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Authentication'
      summary: Signup
      tags:
      - Auth
  /user:
    post:
      consumes:
      - application/json
      description: User Create
      parameters:
      - description: Token with the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: User Input
        in: body
        name: authLogin
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: User Create
      tags:
      - user
swagger: "2.0"